LayoutTests/inspector/runtime/runtime-getProperties-expected.txt

Tests RemoteObject.getProperties.


Running: testSetUp

Running: testGetterAndSetter
{
    enumerable : true
    getter : {
        description : "function () { return 1; }"
        objectId : <string>
        type : "function"
    }
    name : "foo"
    setter : {
        description : "function (value) { }"
        objectId : <string>
        type : "function"
    }
    writable : false
}

Running: testGetterOnly
{
    enumerable : true
    getter : {
        description : "function () { return 1; }"
        objectId : <string>
        type : "function"
    }
    name : "foo"
    setter : {
        description : "undefined"
        objectId : <undefined>
        type : "undefined"
    }
    writable : false
}
